+++ examples/helloworld/src/App.js
@@ -1,27 +1,29 @@
-import logo from './logo.svg';
 import './App.css';
 import '@mintgate/react-mintgate/dist/lib.css'
 import { TPPFormWidget} from '@mintgate/react-mintgate';
 import { TPPLinksList} from '@mintgate/react-mintgate';
 import { TPPLinksGrid} from '@mintgate/react-mintgate';
+import { TPPLinksListUsers} from '@mintgate/react-mintgate';
+import { TPPLinksGridUsers} from '@mintgate/react-mintgate';
 import tokens from './tokens.json';
 function App() {
   return (
     <div className="App">
       <br />
       <h1> Token Gated Form</h1>
-      <TPPFormWidget theme="default"/>
+      <TPPFormWidget 
+        gateTokens={tokens}
+        jwttoken={process.env.JWT || process.env.REACT_APP_JWT} theme="light"/>
       <br />
       <h1> List of token Gated Links Example with $CREATOR</h1>
       <br />
-      <TPPLinksList tokentid="$CREATOR"/>
+      <TPPLinksList tokentid="$CREATOR" theme="light"/>
       <br />
       <h1> Grid of token Gated Links Example with $GATE </h1>
       <br />
-      <TPPLinksGrid tokentid="$GATE" theme="dark"/>
-      <TPPFormWidget 
-        gateTokens={tokens}
-        jwttoken={process.env.JWT || process.env.REACT_APP_JWT} theme="default"/>
+      <TPPLinksGrid base="4" lg="3" md="2" sm="1" xs="1" tokentid="$GATE" theme="light"/>
+      <TPPLinksListUsers jwttoken="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiIxNzciLCJzY29wZXMiOlsid2lkZ2V0Il0sIndpZGdldCI6dHJ1ZSwiaWF0IjoxNjIwNTA0NzYwfQ.DNh6Y8Gbp_ymaRCfMN7Ze9O10Ycq942V4N8KmDPY0sQ" userid="177" theme="light" />
+      <TPPLinksGridUsers jwttoken="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiIxNzciLCJzY29wZXMiOlsid2lkZ2V0Il0sIndpZGdldCI6dHJ1ZSwiaWF0IjoxNjIwNTA0NzYwfQ.DNh6Y8Gbp_ymaRCfMN7Ze9O10Ycq942V4N8KmDPY0sQ" userid="177" theme="light" />
       {/* 
       17 Themes are available: 
       - aqua
