+++ src/components/link-cards/TPPCardWideEmbed.js
@@ -26,12 +26,11 @@ function TPPCardWideEmbed({link, theme}) {
       - valentine  */}
       <div className="mx-4 cursor-pointer">
       <a href={TPP + '/go/' + link.id} target="_blank">
-          <div className=" w-full card bg-base-100 border border-base-300 p-4  shadow-sm hover:shadow-lg transition transform hover:-translate-y-1 motion-reduce:transition-none motion-reduce:transform-none">
+          <div className=" w-full card bg-base-100 border border-base-300 p-4 shadow hover:shadow-none transform hover:scale-95 transition duration-300">
             <div id="header" className="flex flex-wrap sm:flex-nowrap">
-              <img src={link.img} alt="Link Title" className={`${link.img ? 'object-cover md:object-cover w-full min-w-32 sm:w-32 max-h-32 rounded-lg' : 'hidden'}`}/>
               <div id="body" className="flex flex-col mt-6 md:mt-0 ml-5">
-                <h4 id="name" className="card-title text-base-content text-left font-heading text-sm md:text-md font-semibold mb-2">{link.title ? link.title : "Gated Link"}</h4>
-                <p id="job" className="font-body text-xs md:text-sm text-left text-base-content">{link.desc ? link.desc : "A token gated link that is only visitable for the holders of this NFT"}</p>
+                <h2 className="card-title text-base-content text-md text-left font-heading font-semibold">{link.title ? link.title : "Gated Link"}</h2> 
+                <p className="font-body text-base-content text-sm text-left">{link.desc ? link.desc : "A token gated link that is only visitable for the holders of this NFT"}</p> 
               </div>
             </div>
           </div>
@@ -44,4 +43,4 @@ function TPPCardWideEmbed({link, theme}) {
   );
 }
 
-export default TPPCardWideEmbed 
+export default TPPCardWideEmbed
