+++ src/components/token-gating-links/TPPLinksGrid.js
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from "react";
 import Masonry from 'react-masonry-css';
-import TPPCardEmbed from "./TPPCardEmbed.js"
+import TPPCardEmbed from "../link-cards/TPPCardEmbed.js"
 
 const TPP = process.env.NEXT_PUBLIC_TPP_SERVER || `https://mgate.io`;
 
@@ -23,11 +23,11 @@ function TPPLinksGrid(props) {
     console.log('data', tokenData);
 
   const breakpointColumnsObj = {
-    default: 4,
-    1200: 3,
-    800: 2,
-    600: 1,
-    500: 1
+    default: props.base,
+    1200: props.lg,
+    800: props.md,
+    600: props.sm,
+    500: props.xs
   };
 
   return(
