+++ src/components/token-gating-links/TPPLinksList.js
@@ -1,5 +1,5 @@
 import React, {useState, useEffect} from "react";
-import TPPCardWideEmbed from "./TPPCardWideEmbed"
+import TPPCardWideEmbed from "../link-cards/TPPCardWideEmbed"
 
 const TPP = process.env.NEXT_PUBLIC_TPP_SERVER || `https://mgate.io`;
 
@@ -22,29 +22,9 @@ function TPPLinksList(props) {
     console.log('data', tokenData);
 
   return(
-    <div data-theme="light">  {/* 
-      17 Themes are available: 
-      - aqua
-      - black
-      - bumblebee
-      - cupcake
-      - cyberpunk
-      - dark
-      - dracula
-      - fantasy
-      - forest
-      - garden
-      - halloween
-      - light (default)
-      - luxury
-      - pastel
-      - retro
-      - synthwave
-      - valentine  */}
       <div className="space-y-4">
         {tokenData && tokenData.result.map(link =>
-      <TPPCardWideEmbed key={link.id} link={link} />)}
-      </div>
+      <TPPCardWideEmbed key={link.id} link={link} theme={props.theme} />)}
       </div>
   );
 }
